Allow extension to work with Solidus Auth Devise
To load the required files we need to override
SolidusSupport.frontend_available? to return true even if
solidus_frontend is not there.

Since we have no control on view paths order we manually prepended
views of this gem over the ones set by solidus_auth_devise. This
ensures that this extension's views take always precedence.