Create alternate sandbox script which copies Starter FE into app
Goal
----

As a `solidus_starter_frontend` contributor

I want an alternate sandbox script which copies Starter FE into the sandbox
application

So that I have an easy way to confirm if the Starter FE can generate a working
frontend for a Rails app.

Previous implementations
------------------------

* Fix sandbox script https://github.com/nebulab/solidus_starter_frontend/pull/166.

Relevant links
--------------

https://github.com/nebulab/solidus_starter_frontend/pull/167#discussion_r669313064
- discussion on having two strategies for generating the sandbox app: one with
the Starter FE running as an engine and other where Starter FE is generated
into the app.