Update SolidusStarterFrontendGenerator to copy the specs to the storefront app
Goal
----

As a solidus_starter_frontend contributor
I want SolidusStarterFrontendGenerator to copy the test suite to the storefront app
So that I can the run the test suite to verify that the generated frontend is working

Considerations
--------------

We still want the specs to run on the gem. That is, we shouldn't be forced to
generate the app in order to run the test suite there.

Implications
------------

This would mean that SolidusStarterFrontendGenerator would also generate a
starter test suite for the storefronts it would generate.

*Pros*

* Users of SolidusStarterFrontendGenerator can lean on the test suite as they
  change their storefronts.

*Cons*

* We would be including the whole test suite, which may include
  * Low-quality tests
  * Tests that would make more sense in the context of the gem than in the
    context of a generated storefront.
* We could be bloating the user's test suite.

Thoughts
--------

Despite the cons, I think it would be difficult to curate a test suite
specifically for the generated storefront.

Furthermore, considering that we're already allowing the frontend code to be
consumed directly by users, I think that it's okay to afford the test suite the
same leniency we are allowing to the frontend code. As a plus, any improvements
we make later on to the frontend code and its test suite will benefit both the
gem and its consumers using its generated frontends.

Scope
-----

In a succeeding ticket, we can update the CI to run the specs in the generated app.