Fix: guest checkout registration error is not visible
Acceptance
----------

Given allow_guest_checkout is enabled
And I am not logged in
And I have an item in my cart
And I am in the cart page
When I try to check out
Then I should see the checkout registration page
And I should see the Checkout as a Guest form
When I leave the Email field in the form blank
And I click the Continue button of the form
Then I should see an error saying "Email address cannot be blank"

Bug description
---------------

The error message is not visible.

Cause
-----

The registration_error flash message does not have a background color. The background color is needed to contrast with the white color of the flash text.