[BROKEN JS] Copy product form from Solidus Demo
This copies the exact code of almost all the product form code from
Solidus Demo, commit 3c11395ef6bfda0b906f68dbd11ab5ff567da747. The only
exceptions are the partial paths in the product_form template.

Because the code was copied as-is from Solidus Demo, it breaks some
existing functionality in the form. These are:

* The thumbnails are not filtered by variant.
* Selecting a variant does not update the image and its thumbnails.

These issues will be fixed in the next commit.

Other than that, this commit passes CI.