Remove token_order route
Goal
----

As a SolidusStarterFrontend maintainer
I want to remove the token_order route
So that users of the frontend would have a more minimal set of routes to
begin with.

Are there extensions that depend on the token_order_path?
---------------------------------------------------------

`solidus_auth_devise` calls the path helper but only when the gem is
used with the legacy frontend. SolidusStarterFrontend no longer uses
`solidus_auth_devise`'s frontend overrides because
SolidusStarterFrontend has already embedded them to its code.

Other than that, there are no other known calls to `token_order_path`
in Solidus extensions. See

* https://github.com/search?q=org%3Asolidusio+token_order_path&type=code
* https://github.com/search?q=org%3Asolidusio-contrib+token_order_path&type=code