Fix the installation of solidus_auth_devise with v3.2
Adding solidus_auth_devise to the Gemfile was conflicting with it
being added by the solidus installer again unless
`--with-authentication=false` was provided. But the auth extension
is also required by SSF. This forced a workaround in which we skip
the `gem` entry in the Gemfile if another entry for the same gem is
present in any of the following lines. It's ugly but puts a bandaid
that will at least allow the installation to succeed.

Also:
- BundlerContext was used instead of Bundler.
- #run_bundle is not available with the v3.2 generator

Since we're not testing the template when authentication is not
installed this didn't came up in the CI.