Let the sandbox script run the importmap generators
The sandbox script was not running the generators because it was
skipping the initial `bundle install`.

After removing the `--skip-bundle` option it was necessary to run
`rails new` with the correct version of rails without being limited to
the gems in top-level Gemfile.

For this the `unbundled` bash function was added and the Rails version
extracted to a local variable and then used with the (somewhat
obscure) underscore syntax: `<executable> _<gem-version>_`.

A few cleanups were also applied from the solidus_paypal_commerce_platform
sandbox script.