Add explicit responders dependency
With solidusio/solidus#5158, we removed respond_to :html from
Spree::BaseController. That's because responders is no more a
core dependency and BaseController is defined there.

Still, the starter frontend is using it and we can't remove all
occurrences of its usage now, mainly because part of them are
used in the authentication system (it's a devise dependency) and
it takes some time to understand what we can remove/change without
security issues.