Add additional canonical tag test coverage
In a subsequent commit, we want to replace the canonical tag helper. So
having a record of the current behaviour will help us safely
re-implement the functionality.

We changed the spec type to `:feature` because there seems to be a
database persistence issue when running `:system` tests headfully. (i.e.
`ActiveRecord::RecordNotFound` errors for everything!) Seems like a
configuration problem.

Co-authored-by: benjamin wil <benjamin@super.gd>