Allow to skip frontend specs from extensions
When we test extensions with the new starter frontend
we run this template to generate the dummy app. This
template also installs all frontend specs. This leads
to unnecessary long build times, flaky sepcs and simply
wastes a lot of resources.

Since the extension tests it's own behavior in the dummy app
we simply must not duplicate every single spec we also test
in this repo.