Pass in line item to item_info instead of product
This sort of worked before because the only method that is definitely
called on this was name which exists on both line item and product.
Aside from being misleading, this would also error if stock_info was set
to true in the _item_info partial because it would attempt to call
insufficient_stock? on it.